pyproject.toml
[project]
requires-python = ">=3.11"
name = "dummy"
version = "0.1.0"

dependencies = [
    "mcp>=1.13.1",
    "torch", 
    "transformers",
    "datasets",
    "einops",
    "psutil",
    "hf-transfer",
    "fire",
    "einx",
    "openai-agents",
]

license = { file = "LICENSE" }


[project.scripts]
convert_pdf = "convert_pdf_cli:main"

[tool.setuptools]
py-modules = ["convert_pdf_cli"]
